fugu cleanup: compile time warning removals
BZ: 209177

Reviewed-on: https://android.intel.com/214323
Tested-by: Fourdan, Olivier <olivier.fourdan@intel.com>

+++ drivers/external_drivers/intel_media/graphics/rgx/services/3rdparty/intel_drm/pvr_drm.c
@@ -293,7 +293,7 @@ int PVRSRVGetMeminfoPages(void* hMemHandle, int npages, struct page ***pages)
         return -EFAULT;
     }
 
-    kaddr = (uint32_t)minfo.pvCpuVirtAddr;
+    kaddr = (uint32_t)(uintptr_t)minfo.pvCpuVirtAddr;
 
     if ((pglist = kzalloc(npages * sizeof(struct page*),GFP_KERNEL)) == NULL)
     {
@@ -341,7 +341,7 @@ int PVRSRVGetMeminfoPfn(void           *hMemHandle,
         return 0;
     }
 
-    kaddr = (uint32_t)minfo.pvCpuVirtAddr;
+    kaddr = (uint32_t)(uintptr_t)minfo.pvCpuVirtAddr;
 
     if ((pfnlist = kzalloc(npages * sizeof(unsigned long),
                            GFP_KERNEL)) == NULL)
